websocket-server/pwm/requirements.txt
RPi.GPIO==0.6.5